[SPARK-56415][INFRA] Extract shared JIRA utilities into spark_jira_utils.py
