+++ dev/create_jira_and_branch.py
@@ -18,27 +18,11 @@
 #
 
 import argparse
-import os
-import re
 import subprocess
 import sys
 import traceback
 
-try:
-    import jira.client
-
-    JIRA_IMPORTED = True
-except ImportError:
-    JIRA_IMPORTED = False
-
-# ASF JIRA access token
-JIRA_ACCESS_TOKEN = os.environ.get("JIRA_ACCESS_TOKEN")
-JIRA_API_BASE = "https://issues.apache.org/jira"
-
-
-def fail(msg):
-    print(msg)
-    sys.exit(-1)
+from spark_jira_utils import create_jira_issue, fail, get_jira_client
 
 
 def run_cmd(cmd):
@@ -47,49 +31,6 @@ def run_cmd(cmd):
         return subprocess.check_output(cmd).decode("utf-8")
     else:
         return subprocess.check_output(cmd.split(" ")).decode("utf-8")
-
-
-def create_jira_issue(title, parent_jira_id=None, issue_type=None, version=None, component=None):
-    asf_jira = jira.client.JIRA(
-        {"server": JIRA_API_BASE}, token_auth=JIRA_ACCESS_TOKEN, timeout=(3.05, 30)
-    )
-
-    if version:
-        affected_version = version
-    else:
-        versions = asf_jira.project_versions("SPARK")
-        # Consider only x.y.z, unreleased, unarchived versions
-        versions = [
-            x
-            for x in versions
-            if not x.raw["released"]
-            and not x.raw["archived"]
-            and re.match(r"\d+\.\d+\.\d+", x.name)
-        ]
-        versions = sorted(versions, key=lambda x: x.name, reverse=True)
-        affected_version = versions[0].name
-
-    issue_dict = {
-        "project": {"key": "SPARK"},
-        "summary": title,
-        "description": "",
-        "versions": [{"name": affected_version}],
-    }
-
-    if component:
-        issue_dict["components"] = [{"name": component}]
-
-    if parent_jira_id:
-        issue_dict["issuetype"] = {"name": "Sub-task"}
-        issue_dict["parent"] = {"key": parent_jira_id}
-    else:
-        issue_dict["issuetype"] = {"name": issue_type if issue_type else "Improvement"}
-
-    try:
-        new_issue = asf_jira.create_issue(fields=issue_dict)
-        return new_issue.key
-    except Exception as e:
-        fail("Failed to create JIRA issue: %s" % e)
 
 
 def create_and_checkout_branch(jira_id):
@@ -108,10 +49,7 @@ def create_commit(jira_id, title):
         fail("Failed to create commit: %s" % e)
 
 
-def choose_components():
-    asf_jira = jira.client.JIRA(
-        {"server": JIRA_API_BASE}, token_auth=JIRA_ACCESS_TOKEN, timeout=(3.05, 30)
-    )
+def choose_components(asf_jira):
     components = asf_jira.project_components("SPARK")
     components = [c for c in components if not c.raw.get("archived", False)]
     for i, c in enumerate(components):
@@ -130,12 +68,6 @@ def choose_components():
 
 
 def main():
-    if not JIRA_IMPORTED:
-        fail("Could not find jira-python library. Run 'sudo pip3 install jira' to install.")
-
-    if not JIRA_ACCESS_TOKEN:
-        fail("The env-var JIRA_ACCESS_TOKEN is not set.")
-
     parser = argparse.ArgumentParser(description="Create a Spark JIRA issue.")
     parser.add_argument("title", nargs="?", help="Title of the JIRA issue")
     parser.add_argument("-p", "--parent", help="Parent JIRA ID for subtasks")
@@ -148,10 +80,9 @@ def main():
     parser.add_argument("-c", "--component", help="Component for the issue")
     args = parser.parse_args()
 
+    asf_jira = get_jira_client()
+
     if args.parent:
-        asf_jira = jira.client.JIRA(
-            {"server": JIRA_API_BASE}, token_auth=JIRA_ACCESS_TOKEN, timeout=(3.05, 30)
-        )
         parent_issue = asf_jira.issue(args.parent)
         print("Parent issue title: %s" % parent_issue.fields.summary)
         print("Creating a subtask of %s with title: %s" % (args.parent, args.title))
@@ -162,9 +93,11 @@ def main():
         parser.error("the following arguments are required: title")
 
     if not args.component:
-        args.component = choose_components()
+        args.component = choose_components(asf_jira)
 
-    jira_id = create_jira_issue(args.title, args.parent, args.type, args.version, args.component)
+    jira_id = create_jira_issue(asf_jira, args.title, args.component,
+                                parent=args.parent, issue_type=args.type,
+                                version=args.version)
     print("Created JIRA issue: %s" % jira_id)
 
     create_and_checkout_branch(jira_id)
